cmd: split out /e/n/i pretty printer from ifquery.c
