+++ Makefile
@@ -39,8 +39,7 @@ LIBIFUPDOWN_SRC = \
 	libifupdown/lifecycle.c \
 	libifupdown/config-parser.c \
 	libifupdown/config-file.c \
-	libifupdown/compat.c \
-
+	libifupdown/compat.c
 LIBIFUPDOWN_OBJ = ${LIBIFUPDOWN_SRC:.c=.o}
 LIBIFUPDOWN_LIB = libifupdown.a
 
@@ -48,7 +47,8 @@ MULTICALL_SRC = \
 	cmd/multicall.c \
 	cmd/multicall-options.c \
 	cmd/multicall-exec-options.c \
-	cmd/multicall-match-options.c
+	cmd/multicall-match-options.c \
+	cmd/pretty-print-iface.c
 MULTICALL_OBJ = ${MULTICALL_SRC:.c=.o}
 MULTICALL = ifupdown
